Move expression support codes
