+++ packages/delir-core/src/engine/Engine.ts
@@ -19,7 +19,7 @@ import * as ProjectHelper from '../helper/project-helper'
 import { ColorRGB, ColorRGBA } from '../values'
 import AssetProxy from './AssetProxy'
 import DependencyResolver from './DependencyResolver'
-import * as ExpressionContext from './ExpressionContext'
+import * as ExpressionContext from './ExpressionSupport/ExpressionContext'
 import RenderingRequest from './RenderRequest'
 import ClipRenderTask from './Task/ClipRenderTask'
 import EffectRenderTask from './Task/EffectRenderTask'
