+++ packages/delir-core/src/engine/index.ts
@@ -1,5 +1,5 @@
 import * as Renderer from './renderer/index'
 
-export {expressionContextTypeDefinition} from './ExpressionContext'
+export {expressionContextTypeDefinition} from './ExpressionSupport/ExpressionContext'
 export {default as Engine} from './Engine'
 export {Renderer as Renderers}
